+ Bug: Potential issue with Superheavy double-crits and Build Tab Reset button
The equipment list order is important with this.  If you add a later heat sink
before an earlier heat sink, the later heatsink would get stuck, unable to be
removed.

The problem was in UnitUtil.removeCritical.  In the case of two Mounteds in
the same critical slot, it only matched to the first mounted, not the second,
so the order that they get removed in was important.