Fix for printing issue in issue #23.
There was an issue where trying to print from OS X was causing MML to crash. The problem seems to be that the Graphics2D.drawImage() calls in printMekImage() in both the PrintMech and PrintTripod classes were implemented incorrectly.  They included a parameter for color which was causing issues.  Updating these methods to use the correct method call fixes the issue.