Combine like weapon bays to compact weapons list
Warships have lots of weapons, so on the RS bays that are identical across
sides get combined on the sheets.  So, if the FLS and FRS have the same bays
the display will read FLS/FRS for the loc, indicating two bays in each side.

Added code to implement this.