Updated double-line printing criteria
Standard weapons with a lot of damage need to be printed on two lines, to avoid
column overlaps.  The current criteria was if the short range AV was above 100.
Turns out, if the damage drops off with range, and the med AV was < 100, we can
still safely print on one line.