Added two classes to replace unit-specific armor allocation view classes:
ArmorLocationView deals with the number of points allocated to one
location.
ArmorAllocationView positions the location views and shows summary
statistics.

At this point there are combo boxes to set the armor type for the
location if using patchwork armor, but they make the controls too wide
and even moving the armor allocation to a separate panel wouldn't change
the fact that it looks bad. Committing what's been done so far before
changing it.