Mixed tech mechs use internal structure of the designated tech base.
Practically this means that superheavies cannot use Clan structures
(which they shouldn't be able to anyway; SH structure is not the same as
standard, and there is no such thing as Clan superheavy endo-steel).
Also QuadVees will not be able to use IS structure. LAMs also cannot use
Clan IS, but their structure restrictions already cover that.