Changes armor pip layout algorithm to one more suitable to mechs.
Instead of having all the pips positioned in the svg file, the template
file defines regions which MML uses to lay out rows of pips. This is
more flexible as it will allow MML to resize pips if necessary to meet
the required number.

Left the previous method in the file for use with vehicles and aerospace
unit.