Layout changes:
Remove leftover blank on quad and quadvee sheets.
Moved tripod armor diagram drop shadow to right side; the armor diagram
was moved to make room for a possible shield but the drop shadow was not
moved. Moved some of the text that was overlapping the drop shadow.
Fixed center justification of IS value text on quad sheet.