Ignore ammo munition type when printing record sheets.
Some of the ammo names are too long to fit into the space on the
critical table and the actual type of ammo loaded is generally up to the
player(s) rather than something that should be set on the unit. One
exception to this is MML ammo, which has a different number of shots/ton
for SRM and LRM, which should be noted on the sheet. But these are
implemented as separate ammo types rather than submunitions, so are
unaffected by this change.