Only set secondary location for splittable mounts when actually split.
Drag and drop was setting both primary and secondary locations for mechs
even if the primary location had sufficient space, which in turn was
flagging the equipment as split. Though saving and loading would fix it,
this was causing LAM validation to fail when checked during the
construction process.

Fixes #126: LAM split crit weapons erroneous warning