More meaningful setting for armor type combobox.
Units that have patchwork armor will have it set to the armor type of
the first armored location, the same as non-patchwork, so it will have a
more meaningful value if changed away from patchwork.

Also fixed to use first armored location for all units instead of always
using location index zero, which is not armored for vehicles.