Set BA armor tech level before armor type.
The armor type lookup method used by `BattleArmor` prepends "IS" or
"Clan" to the armor name if not already present based on the armor tech
level. This makes it impossible to switch between IS and Clan armor, and
results in all kinds of problems if the selected armor type does not
exist for the previous tech base. Assigning the tech level first
resolves it.

Fixes #305: Mixed IS techbase & Clan BA Fire Resistant