Fix #1230: 0 Max Bay Doors on Large Fixed-Wing Sup
This fix adds support for maximum door counts for Fixed-Wing Support
Vehicles, as well as a count of minimum doors required for all transport
bays, based on TM errata 6.0.

This fix changes the Transport Tab refresh behavior: now, when the
chassis type is changed in the Structure Tab, the Transport Tab will
be completely re-initialized.  This is so that new Fixed Wing Support
Vehicles properly display the door count information (currently it
only displays when loaded from a file or the chache).

NOTE: depends on the same-name PR for MegaMek, which adds the 'minDoors'
field for every Bay type and corrects the Max Door check in
TestAero.java.  Do not pull prior to that PR being pulled.